requirements/base.txt
asn1crypto
certifi
cffi
chardet
cryptography
django-braces
django-crispy-forms
django-debug-toolbar
django-filebrowser
django-grappelli
django-haystack
django-model-utils
ecdsa
elasticsearch
enum34
Fabric
idna
ipaddress
lxml
olefile
paramiko
Pillow
pyasn1
pycparser
pycrypto
pyoai
pytz
requests
Sickle
six
sqlparse
Unipath
urllib3
Whoosh
